+++ crates/commands_robbb/src/checks.rs
