+++ crates/commands_robbb/src/commands/ban.rs
