+++ crates/commands_robbb/src/commands/emojistats.rs
