+++ crates/commands_robbb/src/commands/errors.rs
