+++ crates/commands_robbb/src/commands/help.rs
