+++ crates/commands_robbb/src/commands/info.rs
