+++ crates/commands_robbb/src/commands/mute.rs
