+++ crates/commands_robbb/src/commands/purge.rs
