+++ crates/db_robbb/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod db;
 pub use db::*;
-pub(crate) use shared_robbb::*;
